requirements.txt
# Central Backend Server Dependencies

# FastAPI and Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
argon2-cffi>=23.1.0

# Validation (compatible versions without Rust requirement)
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
email-validator>=2.0.0

# Environment
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0

# Scheduling (for commission calculator)
apscheduler>=3.10.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
